coreboot: Refactor plugin to add metadata to flashrom devices
The coreboot plugin never actually gained the ability to write. As it stands a
coreboot system now adds *two* system-firmware devices (from both flashrom and
coreboot) which isn't ideal.

Just allow flashrom to enumerate quirked devices and add coreboot-specific
metadata as required when the device is registered. If we require some kind
of cbfs parsing then we can do that in FuFlashromDevice->prepare_firmware().